refactor: convert e2e test from standalone script to jest
- Move test logic to test/e2e/conversion.spec.ts (jest describe/it)
- Add jest.e2e.config.js with 30min timeout
- Dockerfile: snapshot consumer-server with devDeps before pruning,
  copy to /consumer-server-e2e in final image for ts-jest
- CI runs: npx jest --config jest.e2e.config.js
- Remove standalone e2e-conversion-test.ts and package.json script
- Revert AB_VERSION back to v48
