+++ consumer-server/package.json
@@ -8,7 +8,6 @@
     "test": "jest --forceExit --detectOpenHandles --coverage --verbose",
     "test-conversion": "node --trace-warnings --abort-on-uncaught-exception --unhandled-rejections=strict dist/test-conversion.js",
     "migrate": "node --trace-warnings --abort-on-uncaught-exception --unhandled-rejections=strict dist/migrate-to-canonical.js",
-    "e2e-conversion-test": "node --trace-warnings --abort-on-uncaught-exception --unhandled-rejections=strict dist/e2e-conversion-test.js",
     "lint:check": "eslint '**/*.{js,ts}'",
     "lint:fix": "eslint '**/*.{js,ts}' --fix"
   },
